Add initial setup files for Astro project with Vscode and Tailwind config
This commit introduces the basic setup files for the project. It includes the Astro's source page (src/pages/index.astro) and VSCode configuration files (launch.json and extensions.json). Additionally, Tailwindcss configuration (tailwind.config.mjs) and other necessary setup files are added to ensure proper functioning and development of the project. A '.gitignore' file is also added to keep track of files that should not be tracked by Git.