Add new `about` page to Astro
This commit introduces a new page, `about`, to the Astro website. The page includes basic meta-tags, a script to manage a light and dark theme based on the user's preference and the actual page layout. The page layout contains a navbar with a search bar and buttons to toggle between light and dark mode, and links to the website's home page and Github repository. The `about` page also introduces a new component, `Icon`, from the `astro-icon/components` module.