Create new `recipes` page with necessary scripts and components
This commit introduces a new `recipes` page including necessary meta-tags, a theme managing script and full page layout. The layout entails a navigation bar with search functionality, a Github link, and a button to toggle theme mode. The `Icon` component from the `astro-icon/components` module is also implemented in this page.