Improve website accessibility and clean up layout
The commit enhances the accessibility of the website by adding alt-text to the hero image. It has also cleaned up the layout CSS styles in MainLayout.astro and made small text changes in a few components to improve readability and understanding. Changes include modifying the path for the hero image SVG, adjusting some ARIA labels for better clarity, and enabling pre-rendering for about.astro page.