Add Beef Tongue Salad recipe and update dependencies
A new recipe for Beef Tongue Salad has been added to the recipes section. This commit also includes important dependency updates, with the @astrojs/check version upgraded from 0.5.5 to 0.5.6 and the astro version from 4.4.3 to 4.4.15. These updates ensure the inclusion of the latest bug fixes and improvements in the project.